moved convonet to other file
